Use location information from terminators to generate line_hint
line_hint get the location from the first instruction in its basic block
that has a location annotation. If only the terminator (a.k.a. last)
instruction in the block has such an annotation, it should be used. It
was not used, just because it is treated specially by the compiler and
is not part of the list of instructions in the block.

Additionally, this diff carries location annotation from c_literal to
b_literal, so that a return of a literal is now properly instrumented by
the line coverage.