Carry accurate line information for (some) switches and pattern matches.
This required associating with the switches the line number of the `case` keyword,
rather than the line number of the first clause as it was doing previously.

It required quite a bit of plumbing to get the annotations in the
right place. Here is the main path:
expr (on #c_case)
-> kmatch
-> match
-> match_varcon
-> match_con
-> build_select
The annotation is carried through, transformed from a list to a map in
kmatch (and dropped if line coverage is not enabled)
, then ends up in the #cg_select which is built by build_select.
There is also a recursive path match_varcon -> match_var -> match

The annotation in #cg_select then gets used in the following path:
match_cg (on #cg_select)
-> select_cg
-> select_val_cg
-> make_cond / make_cond_branch / directly used in #b_switch

This was tested by compiling beam_core_to_ssa itself with
+line_coverage, and checking that it worked without errors, and that the
generated beam file looks sane.