beam_debug_info: Use an extensible format for chunk
At the moment, we only store frame-size and var-mappings
in the DbgB chunk, but we would like to store more information.
This will require to bump the version of the chunk encoding.

In doing this, we would like to make it so that it is easier to
extend this in the future retaining backwards compatibility.

Where today every item is encoded with a `call` instruction,
where the first argument is the frame size and the secod one
it the var mapping, we will now use a `call` instruction with
the first argument being a key (for now "frame size" or "var
mappings"), and the second a payload.

As long as the loader is prepared for missing entries, we
retain backwards compatibility.